apps/yet-another-demo-project-main/auth/requirements.txt
PyJWT
Flask
flask_sqlalchemy
werkzeug
SQLAlchemy